WA-417 Small refactoring
